feat: move files, use actions to handle sync action
