+++ src/scripts/modules/configurations/DockerActionsConstants.js
