+++ src/scripts/modules/configurations/DockerActionsValidityConstants.js
